migrando el código para agregarlos a los scripts de bash de PRSsat

+++ PRSpost/convertir_imgs_irb.py
@@ -2,10 +2,11 @@
 # -*- coding: utf-8 -*-
 
 import sys
+import os.path
 
 from file_to_png import fileToPng
 from shutil      import copyfile
-from funciones   import ymd, getLastFile, copiar_frames
+from funciones   import copiar_frames, makeTimestamp
 
 PATHpng = '/sat/prd-sat/PNGs/'
 
@@ -13,16 +14,13 @@ baseIR  = '/sat/prd-sat/ART_G060x060GG_C060x060/'
 
 meta60  = baseIR + 'meta/'
 
-# lista = open(sys.argv[1], 'r')
-
-lista = open('/sat/PRS/libs/PRS-auto/data/job_imglist_IRB1', 'r')
+# abro el archivo cuya ruta es el primer parametro
+# debe ser el archivo job_imglist_IRB1
+lista = open(sys.argv[1], 'r')
 
 for f in lista:
 
   year, rootname = getYearRoot(f)
-
-  print year
-  print rootname
 
   B02path = baseIR  + 'B02-T2/' + year + '/' + rootname + '.T2'
   B03path = baseIR  + 'B03-T3/' + year + '/' + rootname + '.T3'
@@ -35,26 +33,22 @@ for f in lista:
   fileToPng(B06path, meta60, PATHpng)
 
   # si es el ultimo archivo procesado
-  if f == lista[-1]:
+  # y ya estan procesados los archivos de VIS
+  if f == lista[-1] and os.path.isfile(PATHpng + 'B01-RP/' + year + '/' + rootname + '.png'):
 
-    file = PATHpng + 'B02/' + year + '/' + rootname + '.png'
-    copyfile(file, PATHpng + "BAND_02.png")
+    fileB02 = PATHpng + 'B02/' + year + '/' + rootname + '.png'
+    fileB03 = PATHpng + 'B03/' + year + '/' + rootname + '.png'
+    fileB04 = PATHpng + 'B04/' + year + '/' + rootname + '.png'
+    fileB06 = PATHpng + 'B06/' + year + '/' + rootname + '.png'
 
-    file = PATHpng + 'B03/' + year + '/' + rootname + '.png'
-    copyfile(file, PATHpng + "BAND_03.png")
-
-    file = PATHpng + 'B04/' + year + '/' + rootname + '.png'
-    copyfile(file, PATHpng + "BAND_04.png")
-
-    file = PATHpng + 'B06/' + year + '/' + rootname + '.png'
-    copyfile(file, PATHpng + "BAND_06.png")
+    copyfile(fileB02, PATHpng + "BAND_02.png")
+    copyfile(fileB03, PATHpng + "BAND_03.png")
+    copyfile(fileB04, PATHpng + "BAND_04.png")
+    copyfile(fileB06, PATHpng + "BAND_06.png")
 
     copiar_frames(PATHpng + 'B04/' + year, PATHpng + 'B04/mp4')
 
-    doy       = rootname[8:11]                  # obtengo el doy del rootname
-    hms       = rootname[12:18]                 # obtengo la hora minuto y segundo del rootname
-    month     = ymd(int(year), int(doy))[1]     # obtengo el mes usando la funcion ymd
-    timestamp = year + '.' + str(month).zfill(2) + '.' + str(doy).zfill(3) + '.' + hms
+    timestamp = makeTimestamp(year, rootname)
 
     timestamp_file = open(PATHpng + 'timestamp.html', 'w')
     timestamp_html.write(timestamp)
